refactor: rename ProviderNode.tsx → ServiceNode.tsx, ProviderNodeContent.tsx → ServiceNodeContent.tsx
