+++ frontend/src/components/nodes/ServiceNodeContent.tsx
